requirements.txt
celery==4.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.3.2
PyMySQL==1.0.2
python-decouple==3.4
redis==3.5.3
requests==2.26.0